requirements.txt
cffi==1.14.0
click==7.1.2
cryptography==2.9.2
Flask==1.1.2
Flask-Testing==0.8.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==1.7.1
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
psycopg2==2.8.5
requests~=2.24.0